Rename file, postResetPasswordActions to resetPasswordActions
