+++ src/login/components/LoginApp/ResetPassword/ExtraSecurity.js
@@ -5,12 +5,12 @@ import EduIDButton from "../../../../components/EduIDButton";
 import { useDispatch, useSelector } from "react-redux";
 import ResetPasswordLayout from "./ResetPasswordLayout";
 import PropTypes from "prop-types";
-import { requestPhoneCode, selectExtraSecurity } from "../../../redux/actions/postResetPasswordActions";
+import { requestPhoneCode, selectExtraSecurity } from "../../../redux/actions/resetPasswordActions";
 import ExtraSecurityToken from "../ResetPassword/ExtraSecurityToken";
 import { assertionFromAuthenticator } from "../../../app_utils/helperFunctions/authenticatorAssertion";
 import Splash from "../../../../containers/Splash";
 import { eduidRMAllNotify, eduidNotify } from "../../../../actions/Notifications";
-import { saveLinkCode } from "./../../../redux/actions/postResetPasswordActions";
+import { saveLinkCode } from "../../../redux/actions/resetPasswordActions";
 import { cancelWebauthnAssertion } from "../../../redux/actions/getWebauthnAssertionActions";
 
 const SecurityKeyButton = ({
