+++ src/login/redux/sagas/resetpassword/postResetPasswordSaga.js
@@ -4,7 +4,7 @@ import {
   putCsrfToken
 } from "../../../../sagas/common";
 import postRequest from "../postDataRequest";
-import { postEmailLinkFail } from "../../actions/postResetPasswordActions";
+import { postEmailLinkFail } from "../../actions/resetPasswordActions";
 import { history } from "../../../components/App/App";
 import { countFiveMin, LOCAL_STORAGE_PERSISTED_COUNT_RESEND_LINK, clearCountdown, setLocalStorage } from "../../../components/LoginApp/ResetPassword/CountDownTimer";
 import {
