+++ src/tests/sagas/postVerifyEmailSaga-test.js
@@ -2,7 +2,7 @@ import expect from "expect";
 import { call } from "redux-saga/effects";
 import postRequest from "../../login/redux/sagas/postDataRequest";
 import { useLinkCode } from "../../login/redux/sagas/resetpassword/postVerifyEmailSaga";
-import { postLinkCodeFail } from "../../login/redux/actions/postResetPasswordActions";
+import { postLinkCodeFail } from "../../login/redux/actions/resetPasswordActions";
 
 const fakeState = {
   config: {
